api-learnejo/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.21
alembic==1.12.0
pymysql==1.1.0
python-dotenv==1.0.0
pyjwt==2.12.1
requests==2.31.0
cryptography==41.0.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
